Improve code performance
Previously I was causing a horrible rerender issue whenever the state
changed. Since we're dealing with a 3rd party library that already
handles rerendering the chart, there is no need to do it with React.

Move the graphing to the main App component and let the chart work its
magic